+++ src/features/step_definitions/run_and_view_interviewer_call_history_report.ts
@@ -1,9 +1,7 @@
 import {defineFeature, loadFeature} from "jest-cucumber";
-import "@testing-library/jest-dom";
 import {createMemoryHistory} from "history";
-import {cleanup, screen, waitFor} from "@testing-library/react";
-import ReactDOM from "react-dom";
-import {Router} from "react-router";
+import {cleanup, render, screen, waitFor} from "@testing-library/react";
+import {Router} from 'react-router-dom'
 import App from "../../App";
 
 const feature = loadFeature(
@@ -24,8 +22,7 @@ defineFeature(feature, test => {
 
     test("Run and view interviewer call history report", ({given, when, then}) => {
         given("An interviewer ID and time period (start date and end date) has been specified", async () => {
-            const history = createMemoryHistory();
-            // TODO stuck rendering page... ;o
+            const history = createMemoryHistory()
         });
 
         when("I request information on call history for that interviewer within the time specified period", async () => {
