+++ src/interfaces/index.ts
@@ -1,0 +1,10 @@
+interface ReportData {
+    questionnaire_name: string
+    serial_number: string
+    call_start_time: string
+    dial_secs: string
+    number_of_interviews: string
+    call_result: string
+}
+
+export type {ReportData};
