+++ src/utilities/http/requestPromise.ts
@@ -2,7 +2,6 @@ type PromiseResponse = [number, any];
 
 function requestPromiseJson(method: string, url: string, body: any = null): Promise<PromiseResponse> {
     return new Promise((resolve: (object: PromiseResponse) => void, reject: (error: string) => void) => {
-        console.log(body);
         fetch(url, {
             "method": method,
             "body": body
